some refactoring done
