+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 .env
 .txt
+__pycache__/
